# swagger/api.yaml
swagger: '2.0'
info:
  description: This is a bl-api-index server.
  version: 0.0.1
  title: BlueLens Index API
  contact:
    email: devops@bluehack.net
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'index.stylelens.io'
basePath: /
tags:
  - name: image
    description: image
schemes:
  - http
paths:
  /images:
    post:
      tags:
        - image
      summary: Add a new image
      description: ''
      operationId: addImage
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Image object that needs to be added to the db.
          required: true
          schema:
            $ref: '#/definitions/Image'
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/AddImageResponse'
        400:
          description: Invalid input
  /images/{imageId}:
    get:
      tags:
      - image
      summary: Get a image
      description: ''
      operationId: getImageById
      produces:
      - application/json
      parameters:
      - name: imageId
        in: path
        description: ID of Image to return
        required: true
        type: integer
        format: int64
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Image'
        400:
          description: Invalid input
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  Image:
    type: object
    properties:
      name:
        type: string
      host_url:
        type: string
      host_code:
        type: string
      host_name:
        type: string
      tags:
        type: array
        items:
          type: string
      format:
        type: string
      product_name:
        type: string
      parent_image_raw:
        type: string
      parent_image_mobile:
        type: string
      parent_image_mobile_thumb:
        type: string
      image:
        type: string
        example: "s3://bluelens-style-object/n0100016/59e17359f8375c0e9fabb1c8.jpg"
      class_code:
        type: string
        example: "n0100016"
      product_price:
        type: string
      currency_unit:
        type: string
      product_url:
        type: string
      product_no:
        type: string
      main:
        type: integer
      nation:
        type: string
  AddImageResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
        description: "0: success, 1: fail"
      message:
        type: string
      data:
        type: object
        properties:
          _id:
            type: string
  GetImageResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
        description: "0: success, 1: fail"
      message:
        type: string
      data:
        $ref: '#/definitions/Image'
externalDocs:
  description: Find out more about BlueLens
  url: 'http://bluelens.io'